requirements.txt
click==7.1.2
joblib==1.0.0
nltk==3.5
numpy==1.19.5
pdf2image==1.14.0
Pillow==8.1.0
pytesseract==0.3.7
regex==2020.11.13
tqdm==4.56.0
